requirements.txt
fastapi
uvicorn
python-multipart
numpy<2.0.0
Pillow<11.0.0
scipy
torch
torchvision
opencv-python
ultralytics
mediapipe
facenet-pytorch
scikit-image
